Change config file location + Small fixes (#16)
Hi! I worked on some small changes and I wanted to propose them to you.
Feel free to reject them or to suggest improvements. Here is what I did.

- I didn't particularly like to have the configuration file in
`~/.gptrc` for two reasons: (1) I don't like polluting my home folder
too much and (2) the "rc" that you see for example in "bashrc" stands
for "run commands" and is usually meant to identify a file that can be
executed by a shell, but in our case the file is a yaml configuration
file. So I moved it to `~/.config/gpt-cli/gpt.yml` as it is quite
common.
- I consequently updated the documentation, also adding information on
how to run the program from anywhere.
- I also made a slight change in the way python handles the path to make
it compatible with Windows.
- I added missing arguments in a formatted string.

Thank you for the great project, I hope I can contribute more in the
future!